CORE-387: Update Demo; Handle BlockchainDB tiers; 'cost factor' is a double.
For BTC if the 'pricePerCostFactor' is SAT/KB then the 'costFactor' needs to be a double (since BTC transaction can be <1 KB).  Fix a race on bwm->requestId++.

+++ crypto/BRCryptoFeeBasis.h
@@ -31,7 +31,7 @@ extern "C" {
     extern BRCryptoUnit
     cryptoFeeBasisGetPricePerCostFactorUnit (BRCryptoFeeBasis feeBasis);
 
-    extern uint64_t
+    extern double
     cryptoFeeBasisGetCostFactor (BRCryptoFeeBasis feeBasis);
 
     extern BRCryptoAmount
